requirements.txt
fastapi
uvicorn
jinja2
python-multipart
httpx
pywebview
